+++ rust_crypto/src/random.rs
@@ -1,3 +1,5 @@
+// Copyright (C) 2025 Steel Security Advisors LLC
+// SPDX-License-Identifier: GPL-3.0-or-later
 //! Cryptographically Secure Random Number Generation
 //!
 //! Provides secure random number generation using the OS CSPRNG.